+++ src/Config/Config.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AlibabaCloud\Config;
+
+use AlibabaCloud\Client\Resolver\VersionResolver;
+
+/**
+ * @method static V20190108\ConfigApiResolver v20190108()
+ */
+class Config extends VersionResolver
+{
+}